Рефакторинг кода. Разбил код на модули
